+++ Gruntfile.js
@@ -34,6 +34,7 @@ module.exports = function(grunt) {
               'src/segment-parser.js',
               'src/m3u8/m3u8-parser.js',
               'src/xhr.js',
+              'src/playlist.js',
               'src/playlist-loader.js',
               'node_modules/pkcs7/dist/pkcs7.unpad.js',
               'src/decrypter.js'