+++ src/metadata-stream.js
@@ -37,7 +37,8 @@
           if (tag.data[i] === 0) {
             // parse the text fields
             tag.description = parseUtf8(tag.data, 1, i);
-            tag.value = parseUtf8(tag.data, i + 1, tag.data.length);
+            // do not include the null terminator in the tag value
+            tag.value = parseUtf8(tag.data, i + 1, tag.data.length - 1);
             break;
           }
         }
@@ -171,13 +172,6 @@
                    (tag.data[17] << 16) |
                    (tag.data[18] << 8)  |
                    (tag.data[19]);
-      }
-
-      // adjust the PTS values to align with the video and audio
-      // streams
-      if (this.timestampOffset) {
-        tag.pts -= this.timestampOffset;
-        tag.dts -= this.timestampOffset;
       }
 
       // parse one or more ID3 frames