+++ test/manifest/byteRange.js
@@ -137,5 +137,6 @@
     }
   ],
   "targetDuration": 10,
-  "endList": true
+  "endList": true,
+  "discontinuitySequence": 0
 }