+++ test/manifest/emptyAllowCache.js
@@ -13,5 +13,6 @@
     }
   ],
   "targetDuration": 10,
-  "endList": true
+  "endList": true,
+  "discontinuitySequence": 0
 }