+++ test/manifest/invalidPlaylistType.js
@@ -28,5 +28,6 @@
     }
   ],
   "targetDuration": 10,
-  "endList": true
+  "endList": true,
+  "discontinuitySequence": 0
 }