+++ test/manifest/playlistMediaSequenceHigher.js
@@ -9,5 +9,6 @@
     }
   ],
   "targetDuration": 8,
-  "endList": true
+  "endList": true,
+  "discontinuitySequence": 0
 }